Move logger setup code into binary
