+++ bin/Cargo.toml
@@ -74,6 +74,7 @@ path = "src/named.rs"
 
 [dependencies]
 clap = { version = "3.0", default-features = false, features = ["std", "cargo"] }
+env_logger = "0.9"
 futures = { version = "0.3.5", default-features = false, features = ["std"] }
 log = "0.4"
 rustls = { version = "0.20", optional = true }
@@ -81,7 +82,7 @@ time = "0.3"
 tokio = { version = "1.0", features = ["time"] }
 trust-dns-client = { version = "0.21.1", path = "../crates/client" }
 trust-dns-proto = { version = "0.21.1", path = "../crates/proto" }
-trust-dns-server = { version = "0.21.1", path = "../crates/server", features = ["logger"] }
+trust-dns-server = { version = "0.21.1", path = "../crates/server" }
 
 [dev-dependencies]
 env_logger = "0.9"
